Initial Checkin
This is the initial checkin of my minimal wiki scripts.  There's very little to it. But it's functional.

Make sure you have bottle (bottlepy.org) installed.

Pull this all down and run "wiki.py" on the command line.  It'll run the internal simple wsgi server and start up on port 8080.